requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
cloudinary==1.41.0
python-dotenv==1.0.1
